+++ src/Pororoca.Desktop/Program.cs
@@ -2,8 +2,10 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.ReactiveUI;
+using Pororoca.Desktop.UserData;
 using Pororoca.Desktop.ViewModels;
 using Pororoca.Desktop.Views;
+using Pororoca.Domain.Features.Common;
 
 namespace Pororoca.Desktop;
 
@@ -13,9 +15,19 @@ public static class Program
     // SynchronizationContext-reliant code before AppMain is called: things aren't initialized
     // yet and stuff might break.
     [STAThread]
-    public static void Main(string[] args) =>
-        BuildAvaloniaApp()
-        .StartWithClassicDesktopLifetime(args);
+    public static void Main(string[] args)
+    {
+        PororocaLogger.Instance = new(UserDataManager.GetUserDataFolder());
+        try
+        {
+            BuildAvaloniaApp()
+            .StartWithClassicDesktopLifetime(args);
+        }
+        catch (Exception ex)
+        {
+            PororocaLogger.Instance?.Log(PororocaLogLevel.Fatal, "Program crashed!", ex);
+        }
+    }
 
     // Avalonia configuration, don't remove; also used by visual designer.
     public static AppBuilder BuildAvaloniaApp()